Use new dfej options to avoid making any symbolic links. Add --nocleanup option to not delete directory under /tmp. Notice if gzipped output .inv file exists. Suppress "-nowarn" to jikes if --verbose is specified to daikon.pl. Capture standard error as well as standard output when running commands. Add code to run Ajax.

+++ scripts/daikon.pl
@@ -15,8 +15,10 @@ sub usage() {
 	"  -t, --textfile          Save text of invariants in a .txt file\n",
 	"  -v, --verbose           Display progress messages\n",
 	"  -c, --cleanup           Remove files left over from an interrupted session before starting\n",
+        "      --nocleanup         Do not remove temporary files (for debugging)\n",
 	"  -n, --nogui             Do not start the gui\n",
 	"  -s, --src               Make an archive of the source for later reference\n",
+	"  -a, --ajax              Run Ajax analysis\n",
 	"\n",
 	"Example:\n",
 	"  daikon --output test1 packfoo.MyTestSuite 200\n",
@@ -37,6 +39,7 @@ $TAR_MANIFEST_TAG = $ENV{'TAR_MANIFEST_TAG'} || '-T'; # change to -I for athena 
 $DAIKON_WRAPPER_CLASSPATH = $ENV{'DAIKON_WRAPPER_CLASSPATH'} ||
     $ENV{'CLASSPATH'} ||
     '/g2/users/mernst/java/jdk/jre/lib/rt.jar:/g1/users/mistere/java';
+$AJAX_DIR = $ENV{'AJAX_DIR'} || "ajax";
 
 $cp_lib = $DAIKON_WRAPPER_CLASSPATH;
 $cp_dot = $cp_lib . ':.';
@@ -48,9 +51,13 @@ GetOptions("instrument=s" => \@instrument,
 	   "textfile" => \$textfile,
 	   "verbose" => \$verbose,
 	   "cleanup" => \$cleanup,
+	   "nocleanup" => \$nocleanup,
 	   "nogui" => \$nogui,
+	   "n" => \$nogui,
 	   "src" => \$src,
+	   "ajax" => \$ajax,
 	   ) or usagedie();
+my $nowarn = ($verbose ? "-nowarn" : ""); # for jikes
 
 $runnable = shift @ARGV  or usagedie();
 $runnable_args = join(' ', @ARGV);
@@ -114,7 +121,7 @@ if ($output) {
 	$sym = $sym + 1;
 	while ((length $sym) < 4) { $sym = '0' . $sym }
 	$output = "$prefix-$sym";
-    } while (-x "$output.inv" or -x "$output.src.tar.gz");
+    } while (-x "$output.inv" or -x "$output.inv.gz" or -x "$output.src.tar.gz");
 }
 print "Output will go in $output...\n" if $verbose;
 
@@ -126,7 +133,7 @@ unless (which('jikes')) {
 }
 
 # check the program make sure it starts out with no errors
-$error = system("jikes -classpath $cp_dot -depend -nowrite -nowarn $mainsrc");
+$error = system("jikes -classpath $cp_dot -depend -nowrite $nowarn $mainsrc");
 die ("Fix compiler errors before running daikon") if $error;
 
 # come up with a list of files which we need to care about
@@ -135,7 +142,7 @@ print "Building dependency list...\n" if $verbose;
 die (".u files already exist; try running with --cleanup option") if (find("*.u"));
 
 %interesting = (); # keys are interesting java file names
-system("jikes -classpath $cp_dot -depend -nowrite -nowarn +M $mainsrc") && die ("Unexpected jikes error");
+system("jikes -classpath $cp_dot -depend -nowrite $nowarn +M $mainsrc") && die ("Unexpected jikes error");
 for $fname (find("*.u")) {
     open (U, $fname);
     while ($u = <U>) {
@@ -170,60 +177,97 @@ die("Could not archive source") if $err;
 # setup scratch folders
 print "Preparing to instrument files...\n" if $verbose;
 
-die("daikon-java already exists") if (-e 'daikon-java');
-die("daikon-output already exists") if (-e 'daikon-output');
-
 $working = tmpnam();
 die ("tmp filename already exists") if (-e $working);
 mkdir($working, 0700) or die("Could not create tmp dir 1");
-mkdir("$working/daikon-java", 0700) or die("Could not create tmp dir 2");
-symlink($working, "daikon-output") or die("Could not make symlink 1");
-symlink("$working/daikon-java", "daikon-java") or die("Could not make symlink 2");
+mkdir("$working/daikon-instrumented", 0700) or die("Could not create tmp dir 2");
 
 while (1) {
     # instrument the source files
     print "Instrumenting files...\n" if $verbose;
+    my $files;
     if (@instrument) {
-        $dfejcommand = "dfej -classpath $cp_dot " . join(' ', @instrument);
+      $files = join(' ', @instrument);
     } else {
-        $dfejcommand = "dfej -classpath $cp_dot " . join(' ', sort (keys %interesting));
+      $files = join(' ', sort (keys %interesting));
     }
-    $dfejoutput = `$dfejcommand`;
+    $dfejcommand = "dfej -classpath $cp_dot -instrsourcedir=$working/daikon-instrumented/ -declsfiledir=$working/ -tracefilename=$working/the_dtrace_file.dtrace $files";
+    if ($verbose) {
+      my($cwd) = `pwd`;
+      chomp($cwd);
+      print STDERR "dfej: cd $cwd; $dfejcommand\n";
+    }
+    $dfejoutput = `$dfejcommand 2>&1`;
     $dfejerr = $?;
     last if ($dfejerr);
 
+    # run Ajax
+    if ($ajax) {
+      print "Running Ajax over your java program...\n" if $verbose;
+      my($cwd) = `pwd`;
+      chomp($cwd);
+      chdir($working);
+      @decls = find('*.decls', '.');
+      if (!@decls) {
+	print STDERR "No .decls files found\n";
+	chdir($cwd);
+	last;
+      }
+      my $runnable_dots = $runnable;
+      $runnable_dots =~ s:/:.:g;
+      $ajaxcommand = "java -cp $AJAX_DIR "
+	. "ajax.tools.benchmarks.ComparablePairsDescFileReader "
+        . "-cp $AJAX_DIR/tweaked-classes.zip -ap $cwd "
+	. "$runnable_dots -rewrite " . join(" ", @decls);
+      print "Running $ajaxcommand\n" if $verbose;
+      $ajaxoutput = `$ajaxcommand 2>&1`;
+      $ajaxerr = $?;
+      if ($ajaxerr) {
+	chdir($cwd);
+	last;
+      }
+
+      foreach my $d (@decls) {
+	rename($d, $d . ".bak") || die "Cannot create backup of $d!\n";
+	rename($d . ".ajax", $d) || die "Cannot update $d with $d.ajax!\n";
+      }
+
+      chdir($cwd);
+    }
+
     # compile the instrumented source files
     print "Compiling files...\n" if $verbose;
-    $cp_work = "$working/daikon-java:$cp_lib:.";
-    $jikescommand = "jikes -classpath $cp_work -depend -g -nowarn ";
+    $cp_work = "$working/daikon-instrumented:$cp_lib:.";
+    # $jikescommand = "jikes -classpath $cp_work -depend -g $nowarn ";
+    $jikescommand = "jikes -classpath $cp_work -depend -g ";
 
     $mainsrc_was_instrumented =
         (! @instrument) || (grep {$_ eq $mainsrc} @instrument);
     if ($mainsrc_was_instrumented) {
-        $jikescommand .= "$working/daikon-java/$mainsrc";
+        $jikescommand .= "$working/daikon-instrumented/$mainsrc";
     } else {
         $jikescommand .= "$mainsrc";
     }
 
-    $jikesoutput = `$jikescommand`;
+    $jikesoutput = `$jikescommand 2>&1`;
     $jikeserr = $?;
     last if $jikeserr;
 
     # run the test suite / mainline
     print "Running your java program...\n" if $verbose;
-    $javaoutput = `java -classpath $cp_work $runnable $runnable_args`;
+    $javaoutput = `java -classpath $cp_work $runnable $runnable_args 2>&1`;
     $javaerr = $?;
     last if $javaerr;
 
     # find the results
-    $dtrace = join(' ', find('*.dtrace', 'daikon-output/'));
-    $decls = join(' ', find('*.decls', 'daikon-output/'));
+    $dtrace = join(' ', find('*.dtrace', $working));
+    $decls = join(' ', find('*.decls', $working));
     $outerr = !($dtrace && $decls);
     last if $outerr;
 
     # run modbit-munge
     print "Running modbit-munge...\n" if $verbose;
-    $mboutput = `modbit-munge.pl $dtrace`;
+    $mboutput = `modbit-munge.pl $dtrace 2>&1`;
     $mberr = $?;
     last if $mberr;
 
@@ -242,13 +286,18 @@ while (1) {
     last;
 }
 
-unlink("daikon-java") or die("Could not unlink symlink 2 (daikon-java)");
-unlink("daikon-output") or die("Could not unlink symlink 1 (daikon-output)");
-system("rm -rf $working") && die("Could not remove working dir");
+if (! $nocleanup) {
+  system("rm -rf $working") && die("Could not remove working dir");
+}
 
 if ($dfejerr) {
   print $dfejoutput;
   die("dfej error");
+}
+if ($ajaxerr) {
+  print "ajax command: $ajaxcommand\n";
+  print "ajax output: $ajaxoutput\n";
+  die("ajax error");
 }
 if ($jikeserr) {
   print "jikes command: $jikescommand\n";
@@ -261,10 +310,12 @@ if ($javaerr) {
 }
 if ($outerr) {
   if (! $dtrace) {
-    print "No data trace (.dtrace) file found.\n";
+    print "No data trace (.dtrace) file found"
+      . ($verbose ? " in $working" : "") . ".\n";
   }
   if (! $decls) {
-    print "No declaration (.decls) files found.\n";
+    print "No declaration (.decls) files found"
+      . ($verbose ? " in $working" : "") . ".\n";
   }
   die("missing output files");
 }
